// frontend/src/app/core/api/model/subject-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SubjectRequest { 
    code: string;
    name: string;
    description?: string;
    theoryHours: number;
    practicalHours: number;
    defaultCoefficient?: number;
    pdfFilePath?: string;
    specialtyId: number;
}